requirements.txt
nemoguardrails
ollama
langchain_openai
guardrails-ai
python-dotenv